rename sample package to report
