+++ andhow-core/src/main/java/org/yarnandtail/andhow/report/JndiLoaderSamplePrinter.java
@@ -1,4 +1,4 @@
-package org.yarnandtail.andhow.sample;
+package org.yarnandtail.andhow.report;
 
 import org.yarnandtail.andhow.internal.StaticPropertyConfigurationInternal;
 import org.yarnandtail.andhow.AndHow;
@@ -9,10 +9,11 @@ import org.yarnandtail.andhow.util.TextUtil;
  *
  * @author ericeverman
  */
-public class PropFileLoaderSamplePrinter extends BaseSamplePrinter implements SamplePrinter {
-	protected PrintFormat format = new PropFileFormat();
+public class JndiLoaderSamplePrinter extends BaseSamplePrinter implements SamplePrinter {
+	protected PrintFormat format;
 	
-	public PropFileLoaderSamplePrinter() {
+	public JndiLoaderSamplePrinter() {
+		format = new JndiFileFormat();
 	}
 	
 	@Override
@@ -20,28 +21,29 @@ public class PropFileLoaderSamplePrinter extends BaseSamplePrinter implements Sa
 		return format;
 	}
 	
-	
 	@Override
 	public TextBlock getSampleFileStart() {
-		return null;
+		TextBlock tb = new TextBlock(false, false);
+		tb.addLine("<Context>");
+		return tb;
 	}
 	
 	@Override
 	public TextBlock getSampleStartComment(StaticPropertyConfigurationInternal definition) {
 		TextBlock tb = new TextBlock(true, true);
 		tb.addHR();
-		tb.addLine("Sample properties file generated by " + AndHow.ANDHOW_NAME);
+		tb.addLine("Sample JNDI config file generated by " +  AndHow.ANDHOW_NAME);
 		tb.addLine(AndHow.ANDHOW_TAG_LINE + "  -  " + AndHow.ANDHOW_URL);
-		tb.addLine("Note: "+ definition.getNamingStrategy().getNameMatchingDescription());
+		tb.addLine("This sample uses Tomcat syntax. JNDI configuration for other servers will be similar.");
 		tb.addHR();
 		return tb;
-		
 	}
 	
 	@Override
 	public String getInAliaseString(StaticPropertyConfigurationInternal definition, EffectiveName name) {
-		return name.getActualName();
+		return definition.getNamingStrategy().getUriName(name.getActualName());
 	}
+	
 	
 	@Override
 	public TextBlock getActualProperty(StaticPropertyConfigurationInternal definition, 
@@ -49,32 +51,38 @@ public class PropFileLoaderSamplePrinter extends BaseSamplePrinter implements Sa
 		
 		TextBlock tb = new TextBlock(false, false);
 		
-		String propCanonName = group.getCanonicalName(prop);
+		String propCanonName = definition.getNamingStrategy().getUriName(group.getCanonicalName(prop));
+		String type = prop.getValueType().getDestinationType().getCanonicalName();
+
 		
 		if (prop.getDefaultValue() != null) {
 			tb.addLine(
-					TextUtil.format("{} = {}", 
+					TextUtil.format("<Environment name=\"{}\" value=\"{}\" type=\"{}\" override=\"false\"/>", 
 						propCanonName, 
-						prop.getDefaultValue())
+						TextUtil.escapeXml(prop.getDefaultValue().toString()),
+						type)
 			);
 		} else {
 			tb.addLine(
-					TextUtil.format("{} = [{}]", 
+					TextUtil.format("<Environment name=\"{}\" value=\"[{}]\" type=\"{}\" override=\"false\"/>", 
 						propCanonName, 
-						prop.getValueType().getDestinationType().getSimpleName())
+						prop.getValueType().getDestinationType().getSimpleName(),
+						type)
 			);
 		}
 		
 		return tb;
 	}
-	
-	@Override
-	public TextBlock getSampleFileEnd() {
-		return null;
-	}
 
 	@Override
+	public TextBlock getSampleFileEnd() {
+		TextBlock tb = new TextBlock(false, false);
+		tb.addLine("</Context>");
+		return tb;
+	}
+	
+	@Override
 	public String getSampleFileExtension() {
-		return "properties";
+		return "xml";
 	}
 }
