+++ andhow-core/src/main/java/org/yarnandtail/andhow/report/PrintFormat.java
@@ -1,4 +1,4 @@
-package org.yarnandtail.andhow.sample;
+package org.yarnandtail.andhow.report;
 
 import org.yarnandtail.andhow.util.TextUtil;
