+++ andhow-core/src/main/java/org/yarnandtail/andhow/report/PropFileLoaderSamplePrinter.java
@@ -1,4 +1,4 @@
-package org.yarnandtail.andhow.sample;
+package org.yarnandtail.andhow.report;
 
 import org.yarnandtail.andhow.internal.StaticPropertyConfigurationInternal;
 import org.yarnandtail.andhow.AndHow;
@@ -9,11 +9,10 @@ import org.yarnandtail.andhow.util.TextUtil;
  *
  * @author ericeverman
  */
-public class JndiLoaderSamplePrinter extends BaseSamplePrinter implements SamplePrinter {
-	protected PrintFormat format;
+public class PropFileLoaderSamplePrinter extends BaseSamplePrinter implements SamplePrinter {
+	protected PrintFormat format = new PropFileFormat();
 	
-	public JndiLoaderSamplePrinter() {
-		format = new JndiFileFormat();
+	public PropFileLoaderSamplePrinter() {
 	}
 	
 	@Override
@@ -21,29 +20,28 @@ public class JndiLoaderSamplePrinter extends BaseSamplePrinter implements Sample
 		return format;
 	}
 	
+	
 	@Override
 	public TextBlock getSampleFileStart() {
-		TextBlock tb = new TextBlock(false, false);
-		tb.addLine("<Context>");
-		return tb;
+		return null;
 	}
 	
 	@Override
 	public TextBlock getSampleStartComment(StaticPropertyConfigurationInternal definition) {
 		TextBlock tb = new TextBlock(true, true);
 		tb.addHR();
-		tb.addLine("Sample JNDI config file generated by " +  AndHow.ANDHOW_NAME);
+		tb.addLine("Sample properties file generated by " + AndHow.ANDHOW_NAME);
 		tb.addLine(AndHow.ANDHOW_TAG_LINE + "  -  " + AndHow.ANDHOW_URL);
-		tb.addLine("This sample uses Tomcat syntax. JNDI configuration for other servers will be similar.");
+		tb.addLine("Note: "+ definition.getNamingStrategy().getNameMatchingDescription());
 		tb.addHR();
 		return tb;
+		
 	}
 	
 	@Override
 	public String getInAliaseString(StaticPropertyConfigurationInternal definition, EffectiveName name) {
-		return definition.getNamingStrategy().getUriName(name.getActualName());
+		return name.getActualName();
 	}
-	
 	
 	@Override
 	public TextBlock getActualProperty(StaticPropertyConfigurationInternal definition, 
@@ -51,38 +49,32 @@ public class JndiLoaderSamplePrinter extends BaseSamplePrinter implements Sample
 		
 		TextBlock tb = new TextBlock(false, false);
 		
-		String propCanonName = definition.getNamingStrategy().getUriName(group.getCanonicalName(prop));
-		String type = prop.getValueType().getDestinationType().getCanonicalName();
-
+		String propCanonName = group.getCanonicalName(prop);
 		
 		if (prop.getDefaultValue() != null) {
 			tb.addLine(
-					TextUtil.format("<Environment name=\"{}\" value=\"{}\" type=\"{}\" override=\"false\"/>", 
+					TextUtil.format("{} = {}", 
 						propCanonName, 
-						TextUtil.escapeXml(prop.getDefaultValue().toString()),
-						type)
+						prop.getDefaultValue())
 			);
 		} else {
 			tb.addLine(
-					TextUtil.format("<Environment name=\"{}\" value=\"[{}]\" type=\"{}\" override=\"false\"/>", 
+					TextUtil.format("{} = [{}]", 
 						propCanonName, 
-						prop.getValueType().getDestinationType().getSimpleName(),
-						type)
+						prop.getValueType().getDestinationType().getSimpleName())
 			);
 		}
 		
 		return tb;
 	}
-
-	@Override
-	public TextBlock getSampleFileEnd() {
-		TextBlock tb = new TextBlock(false, false);
-		tb.addLine("</Context>");
-		return tb;
-	}
 	
 	@Override
+	public TextBlock getSampleFileEnd() {
+		return null;
+	}
+
+	@Override
 	public String getSampleFileExtension() {
-		return "xml";
+		return "properties";
 	}
 }
