+++ andhow-core/src/main/java/org/yarnandtail/andhow/report/TextLine.java
@@ -1,4 +1,4 @@
-package org.yarnandtail.andhow.sample;
+package org.yarnandtail.andhow.report;
 
 import java.util.ArrayList;
 import java.util.List;
