+++ andhow-core/src/test/java/org/yarnandtail/andhow/report/PropFileLoaderSamplePrinterTest.java
@@ -1,6 +1,6 @@
 /*
  */
-package org.yarnandtail.andhow.sample;
+package org.yarnandtail.andhow.report;
 
 import java.io.UnsupportedEncodingException;
 import org.junit.Before;
@@ -19,7 +19,7 @@ import static org.junit.Assert.*;
  *
  * @author ericeverman
  */
-public class JndiLoaderSamplePrinterTest {
+public class PropFileLoaderSamplePrinterTest {
 	
 	StaticPropertyConfigurationMutable config;
 	GroupProxyMutable groupProxy1;
@@ -36,8 +36,8 @@ public class JndiLoaderSamplePrinterTest {
 		config = new StaticPropertyConfigurationMutable(new CaseInsensitiveNaming());
 		
 		groupProxy1 = new GroupProxyMutable(
-				JndiLoaderSamplePrinterTest.Config.class.getCanonicalName(),
-				JndiLoaderSamplePrinterTest.class.getCanonicalName() + "$Config"
+				PropFileLoaderSamplePrinterTest.Config.class.getCanonicalName(),
+				PropFileLoaderSamplePrinterTest.class.getCanonicalName() + "$Config"
 		);
 		groupProxy1.addProperty(new NameAndProperty("MY_PROP1", Config.MY_PROP1));
 		groupProxy1.addProperty(new NameAndProperty("MY_PROP2", Config.MY_PROP2));
@@ -52,7 +52,7 @@ public class JndiLoaderSamplePrinterTest {
 	public void generalTest() throws UnsupportedEncodingException {
 		
 		TestPrintStream out = new TestPrintStream();
-		JndiLoaderSamplePrinter printer = new JndiLoaderSamplePrinter();
+		PropFileLoaderSamplePrinter printer = new PropFileLoaderSamplePrinter();
 		String[] lines;	//The output line array
 		
 		
@@ -63,12 +63,12 @@ public class JndiLoaderSamplePrinterTest {
 		//System.out.println(out.getTextAsString());
 		lines = out.getTextAsLines();
 		assertEquals(6, lines.length);
-		assertEquals("<Context>", lines[0]);
-		assertEquals("<!-- " + TextUtil.repeat("- ", 45), lines[1]);
-		assertEquals("Sample JNDI config file generated by AndHow!", lines[2]);
-		assertEquals("strong.simple.valid.AppConfiguration  -  https://github.com/eeverman/andhow", lines[3]);
-		assertEquals("This sample uses Tomcat syntax. JNDI configuration for other servers will be similar.", lines[4]);
-		assertEquals(TextUtil.repeat("- ", 45) + " -->", lines[5]);
+		assertEquals("# " + TextUtil.repeat("##", 45), lines[0]);
+		assertEquals("# Sample properties file generated by AndHow!", lines[1]);
+		assertEquals("# strong.simple.valid.AppConfiguration  -  https://github.com/eeverman/andhow", lines[2]);
+		assertEquals("# Note: When reading property names, matching is done in a case insensitive way, so 'Bob'", lines[3]);
+		assertEquals("# 	would match 'bOB'.", lines[4]);
+		assertEquals("# " + TextUtil.repeat("##", 45), lines[5]);
 		
 		//Print group header
 		out.reset();
@@ -77,8 +77,9 @@ public class JndiLoaderSamplePrinterTest {
 		//System.out.println(out.getTextAsString());
 		lines = out.getTextAsLines();
 		assertEquals(2, lines.length);
-		assertEquals("<!-- " + TextUtil.repeat("- ", 45), lines[0]);
-		assertEquals("Property Group " + JndiLoaderSamplePrinterTest.Config.class.getCanonicalName() + " -->",
+		assertEquals("# " + TextUtil.repeat("##", 45), lines[0]);
+		assertEquals(
+				"# Property Group " + PropFileLoaderSamplePrinterTest.Config.class.getCanonicalName(),
 				lines[1]);
 		
 		//Print MY_PROP1
@@ -88,12 +89,11 @@ public class JndiLoaderSamplePrinterTest {
 		//System.out.println(out.getTextAsString());
 		lines = out.getTextAsLines();
 		assertEquals(3, lines.length);
-		assertEquals("<!-- ", lines[0]);
-		assertEquals("MY_PROP1 (Integer) -->", lines[1]);
-		assertEquals("<Environment name=\"" +
-				JndiLoaderSamplePrinterTest.Config.class.getCanonicalName().replace(".", "/")
-				+ "/MY_PROP1\" value=\"[Integer]\" "
-				+ "type=\"java.lang.Integer\" override=\"false\"/>",
+		assertEquals("# ", lines[0]);
+		assertEquals("# MY_PROP1 (Integer)", lines[1]);
+		assertEquals(
+				PropFileLoaderSamplePrinterTest.Config.class.getCanonicalName() +
+						".MY_PROP1 = [Integer]",
 				lines[2]);
 		
 		//Print MY_PROP2
@@ -103,18 +103,17 @@ public class JndiLoaderSamplePrinterTest {
 		//System.out.println(out.getTextAsString());
 		lines = out.getTextAsLines();
 		assertEquals(9, lines.length);
-		assertEquals("<!-- ", lines[0]);
-		assertEquals("MY_PROP2 (String) NON-NULL - mp description", lines[1]);
-		assertEquals("Recognized aliases: mp2, mp2_alias2", lines[2]);
-		assertEquals("Default Value: La la la", lines[3]);
-		assertEquals("Long text on how to use the property", lines[4]);
-		assertEquals("The property value must:", lines[5]);
-		assertEquals("- start with 'La'", lines[6]);
-		assertEquals("- end with 'la' -->", lines[7]);
-		assertEquals("<Environment name=\"" +
-				JndiLoaderSamplePrinterTest.Config.class.getCanonicalName().replace(".", "/")
-				+ "/MY_PROP2\" value=\"La la la\" "
-				+ "type=\"java.lang.String\" override=\"false\"/>",
+		assertEquals("# ", lines[0]);
+		assertEquals("# MY_PROP2 (String) NON-NULL - mp description", lines[1]);
+		assertEquals("# Recognized aliases: mp2, mp2_alias2", lines[2]);
+		assertEquals("# Default Value: La la la", lines[3]);
+		assertEquals("# Long text on how to use the property", lines[4]);
+		assertEquals("# The property value must:", lines[5]);
+		assertEquals("# - start with 'La'", lines[6]);
+		assertEquals("# - end with 'la'", lines[7]);
+		assertEquals(
+				PropFileLoaderSamplePrinterTest.Config.class.getCanonicalName() +
+						".MY_PROP2 = La la la",
 				lines[8]);
 		
 		//Print group closing (should be empty line)
@@ -133,7 +132,7 @@ public class JndiLoaderSamplePrinterTest {
 		//System.out.println(out.getTextAsString());
 		lines = out.getTextAsLines();
 		assertEquals(1, lines.length);
-		assertEquals("</Context>", lines[0]);
+		assertEquals("", lines[0]);
 	}
 
 	/**
