+++ andhow-core/src/test/java/org/yarnandtail/andhow/report/TestPrintStream.java
@@ -1,4 +1,4 @@
-package org.yarnandtail.andhow.sample;
+package org.yarnandtail.andhow.report;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
